@@ -1,6 +1,6 @@
 class Admin::AdminController < ApplicationController
   layout "admin"
-  before_filter :authenticate_user!
+  before_action :authenticate_user!
 
   def current_ability
     @current_ability ||= AdminAbility.new(current_user)
