@@ -1,5 +1,5 @@
 class SearchController < ApplicationController
-  before_filter :authenticate_user!
+  before_action :authenticate_user!
 
   def indicators
   end
