@@ -1,4 +1,4 @@
-class AddConfirmableToDevise < ActiveRecord::Migration
+class AddConfirmableToDevise < ActiveRecord::Migration[4.2]
   def change
   end
 end
