@@ -1,4 +1,4 @@
-class ChangeElementTypeToKind < ActiveRecord::Migration
+class ChangeElementTypeToKind < ActiveRecord::Migration[4.2]
   def change
     add_column :elements, :kind, :string
     remove_column :elements, :type
