@@ -1,4 +1,4 @@
-class AddDeepdiveToSamples < ActiveRecord::Migration
+class AddDeepdiveToSamples < ActiveRecord::Migration[4.2]
   def change
     add_column :samples, :deepdive, :text
   end
